+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
 ## v2.5.0
-- **[feat]** - `SelectMultiple` component
+- **[feat]** - `AutocompleteFilter` component
+- **[chore]** - upgrade storybook to v8.3.6