+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
 ## v2.5.0
-- **[feat]** - `AutocompleteFilter` component
+- **[feat]** - Add new component `AutocompleteFilter`
 - **[chore]** - upgrade storybook to v8.3.6